feat: compare tip to prevent race conditions in event queue
